refactor: update import path for SearchBar component in explore-search and explore-redirect
